Add CustomTextformfield widget and update height/width components to use double type

+++ lib/utils/components/height.dart
@@ -3,7 +3,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class H extends StatelessWidget {
   const H({required this.h, super.key});
-  final int h;
+  final double h;
   @override
   Widget build(BuildContext context) {
     return SizedBox(height: h.h);
