+++ lib/utils/components/width.dart
@@ -3,7 +3,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class W extends StatelessWidget {
   const W({required this.w, super.key});
-  final int w;
+  final double w;
   @override
   Widget build(BuildContext context) {
     return SizedBox(width: w.w);
